+++ build.gradle
@@ -35,6 +35,7 @@ dependencies {
 	implementation 'org.webjars:bootstrap:5.3.7'
 	implementation 'org.webjars:jquery:3.7.1'
 	implementation 'org.thymeleaf.extras:thymeleaf-extras-springsecurity6'
+	implementation 'org.springframework.boot:spring-boot-starter-aop'
 	compileOnly 'org.projectlombok:lombok'
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
 	runtimeOnly 'com.mysql:mysql-connector-j'
