+++ src/main/java/com/example/demo/service/SignupService.java
@@ -2,47 +2,25 @@ package com.example.demo.service;
 
 import java.util.Optional;
 
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-import com.example.demo.constant.AuthorityKind;
 import com.example.demo.entity.UserInfo;
 import com.example.demo.form.SignupForm;
-import com.example.demo.repository.UserInfoRepository;
-
-import lombok.RequiredArgsConstructor;
 
 /**
- * ユーザ登録画面 Service
+ * ユーザ登録画面 Serviceインターフェース
  */
-@Service
-@RequiredArgsConstructor
-public class SignupService {
+public interface SignupService {
 
-	private final UserInfoRepository repository;
-	
-	/** PasswordEncoder */
-	private final PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
-		
 	/**
-	 * ユーザ情報テーブル 新規登録
+	 * 画面の入力情報を元にユーザ情報テーブルの新規登録を行います。
+	 * 
+	 * <p>ただし、以下のテーブル項目はこの限りではありません。</p>
+	 * <ul>
+	 * <li>パスワード：画面で入力したパスワードがハッシュ化され登録されます。</li>
+	 * <li>権限：常に「商品情報の確認が可能」のコード値が登録されます。</li>
+	 * </ul>
+	 * 
 	 * @param form 入力情報
 	 * @return 登録情報（ユーザ情報Entity、既に同じユーザIDで登録がある場合はEmpty
 	 */
-	public Optional<UserInfo> registUserInfo(SignupForm form){
-		Optional<UserInfo> userInfoExisted = repository.findById(form.getLoginId());
-		if (userInfoExisted.isPresent()) {
-			return Optional.empty();
-		}
-		
-		UserInfo userInfo = new UserInfo();
-		userInfo.setLoginId(form.getLoginId());
-		userInfo.setPassword(form.getPassword());
-		String encodedPassword = passwordEncoder.encode(form.getPassword()); //ハッシュ化パスワードの生成
-		userInfo.setPassword(encodedPassword);
-		userInfo.setAuthority(AuthorityKind.ITEM_WATCHER.getAuthorityKind());
-
-		return Optional.of(repository.save(userInfo));
-	}
+	public Optional<UserInfo> registUserInfo(SignupForm form);
 }
