+++ src/main/resources/application.properties
@@ -10,3 +10,9 @@ spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 security.locking-border-count=3
 #アカウントロックの継続時間
 security.locking-time=1
+
+# ログレベル
+logging.level.root=INFO
+logging.level.com.example.demo=DEBUG
+# ログ出力先
+logging.file.name=C:/work/log/debug.log
